moved rollup.config to seperate folder, generate ES, UMD, and IIFE formats
