+++ docs/.vuepress/enhanceApp.js
@@ -1,12 +1,13 @@
 import 'intersection-observer'
 
-import VueSnap from '../../dist/vue-snap.min'
+import VueSnap from '../../dist/vue-snap.umd'
 import lazy from '../../src/directives/lazy'
 
 import VCarousel from '../components/Carousel.vue'
 
 export default ({ Vue }) => {
   Vue.use(VueSnap)
+
   Vue.component('VCarousel', VCarousel)
   Vue.directive('lazy', lazy)
 }
