+++ src/entry.js
@@ -1,0 +1,15 @@
+import './polyfills'
+
+import Carousel from './components/Carousel.vue'
+import Slide from './components/Slide.vue'
+
+const install = Vue => {
+  Vue.component('carousel', Carousel)
+  Vue.component('slide', Slide)
+}
+
+export default {
+  install
+}
+
+export { Carousel, Slide }
